pyproject.toml
[project]
name = "cst-150-project-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "mysql-connector-python>=9.4.0",
]
